api/requirements.txt
fastapi
uvicorn
tensorflow==2.10.0
numpy
pillow
opencv-python
pytest
python-multipart
